backend/requirements.txt
fastapi>=0.95.2
uvicorn>=0.22.0
python-multipart>=0.0.6
httpx>=0.23.0
pydantic>=1.10.0,<2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
tenacity>=8.2.3
structlog>=24.1.0
python-decouple>=3.8
supabase>=1.0.3
